feat: move getting started to drawer
