+++ src/lib/state/db.ts
@@ -133,6 +133,7 @@ window.store = create((set) => {
     combatBuffsDrawerOpen: false,
     enemyConfigurationsDrawerOpen: false,
     settingsDrawerOpen: false,
+    gettingStartedDrawerOpen: false,
     permutations: 0,
     permutationsResults: 0,
     permutationsSearched: 0,
@@ -219,6 +220,7 @@ window.store = create((set) => {
     setCombatBuffsDrawerOpen: (x) => set(() => ({ combatBuffsDrawerOpen: x })),
     setEnemyConfigurationsDrawerOpen: (x) => set(() => ({ enemyConfigurationsDrawerOpen: x })),
     setSettingsDrawerOpen: (x) => set(() => ({ settingsDrawerOpen: x })),
+    setGettingStartedDrawerOpen: (x) => set(() => ({ gettingStartedDrawerOpen: x })),
     setOptimizerTabFocusCharacter: (characterId) => set(() => ({ optimizerTabFocusCharacter: characterId })),
     setCharacterTabFocusCharacter: (characterId) => set(() => ({ characterTabFocusCharacter: characterId })),
     setScoringAlgorithmFocusCharacter: (characterId) => set(() => ({ scoringAlgorithmFocusCharacter: characterId })),
